// src/app/Models/userAccess/IUserAccess.ts
export interface UserAccess {
    uid: string;
    isAdmin: string;
    canShareFeed: string;
    canConnectPeers: string;
    canChat: string;
    canSubmitApplications: string;
    canReferUsers: string;
    salesTally: string[];
    collectionsTarget: string[];
    canViewUserDetailsPOPI: string;
    brandAffiliateChoice: string[];
 }